old-tests/submission/W3C/domhtml/HTMLIFrameElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/dimension-attributes.html#dom-dim-height" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe height="80" src="/support/html/sample.html" width="550">IFRAME1</iframe>


<script>
var domtest = async_test("The height attribute specifies the frame height.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vheight;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vheight = testNode.height;

      assert_equals(vheight,"80");
       
    domtest.done();
}));

</script>
<p>The height attribute specifies the frame height. 

    Retrieve the height attribute of the first IFRAME element and examine
    it's value.</p>
<div id="log"></div>
</body></html>
